Refactor update deprecated keywords (#176)
* Replace `output(Control|Interval)` with `write(Control|Interval)`

`outputControl` -> `writeControl`
`outputInterval` -> `writeInterval`

These keywords have been deprecated in OpenFOAM.COM since
[this commit](https://develop.openfoam.com/Development/openfoam/-/commit/c983670c91d7c5ae613b40b2ebacc8cdfa1874de)
in 2016 (v1612).

They caused the following warnings in OpenFOAM.COM:

```console
--> FOAM Warning :
    From void Foam::timeControl::read(const dictionary &)
    in file db/functionObjects/timeControl/timeControl.C at line 131
    Reading "system/controlDict/functions/forces" at line 66 to 77
    Found deprecated 'outputControl'
    Use 'writeControl' with 'writeInterval'
    This keyword is 102 months old.
```

* Replace `convertToMeters` with `scale`

This keyword was deprecated in OpenFOAM.COM, and its replacement,
`scale`, is supported by all forks.

It caused the following warnings in OpenFOAM.COM:

```console
Creating block mesh from "system/blockMeshDict"
--> FOAM IOWarning :
    Found [v1012] 'convertToMeters' entry instead of 'scale' in dictionary "system/blockMeshDict"

    This keyword is 168 months old.
```

* Replace `functionObjectLibs` with `libs`

NOTE: Some replacements were made in the `src/` directory and in
example comments.

This keyword caused the following warning in OpenFOAM.COM:

```console
--> FOAM IOWarning :
    Found [v1612] 'functionObjectLibs' entry instead of 'libs' in dictionary "system/controlDict/functions/forces"

    This keyword is 96 months old.
```

* Update case scripts to replace project-dependent keywords

The keywords `basePoint` and `normalVector`, which are used in
`system/mirrorMeshDict,` have been deprecated in OpenFOAM-* in favor of
`point` and `normal`, respectively.

These new keywords are not supported in foam-extend; therefore, a manual
conversion was required to avoid the following runtime warnings in
OpenFOAM-*:

```console
Reading mirrorMeshDict

--> FOAM IOWarning :
    Found [v1612] 'basePoint' entry instead of 'point' in dictionary "system/mirrorMeshDict/pointAndNormalDict"

    This keyword is 96 months old.

--> FOAM IOWarning :
    Found [v1612] 'normalVector' entry instead of 'normal' in dictionary "system/mirrorMeshDict/pointAndNormalDict"

    This keyword is 96 months old.
```

Additionally, since they are used only in a single case and in one file,
the replacement is handled in the case's `Allrun` and `Allclean` scripts
to allow conversion back and forth between foam-extend and OpenFOAM-*.

* solids4Foam::convertCaseFormat: check for point/normal in mirrorMeshDict

* solids4FoamScripts.sh: update echos for point/normal replacements

---------

Co-authored-by: Philip Cardiff <philip.cardiff@ucd.ie>